pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "flash512-vanguard"
version = "2.0.0.post1"
authors = [
  { name="erabytse", email="support@fbfconsulting.org" },
]
description = "Professional Secure Hashing & Encryption Library - AES-256-GCM Backend"
readme = "README.md"
requires-python = ">=3.8"
license = "Apache-2.0"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "Topic :: Security :: Cryptography",
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
]
dependencies = [
    "cryptography>=42.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "hypothesis>=6.0",
    "black>=24.0",
    "mypy>=1.0",
]

[tool.setuptools.packages.find]
include = ["flash512*"]
exclude = ["tests*", "menv*"]

[tool.setuptools.package-data]
flash512 = ["py.typed"]  

[project.urls]
"Homepage" = "https://github.com/erabytse/flash512-vanguard"
"Bug Tracker" = "https://github.com/erabytse/flash512-vanguard/issues"